refactor: remove Gemini flash dynamic-thinking model as its enabled by default
Also splits up the providers into individual files for better
maintenance.
