+++ runner/codegen/ai-sdk/anthropic_thinking_patch.ts
